pyproject.toml
[project]
name = "simple-query-responder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "sehrish-hub", email = "sehrishshafiq345@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.60.2",
    "python-dotenv>=1.0.1",
]

[project.scripts]
simple-query-responder = "simple_query_responder:main"
hello = "simple_query_responder.main:kickoff"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
